2-str_concat.c

+++ 0x0B-malloc_free/2-str_concat.c
@@ -20,10 +20,10 @@ concat = (char *)malloc(sizeof(char) * (strlen(s2) + strlen(s1) + 1));
 if (!concat)
 return (NULL);
 
-for (i = 0; i < strlen(s1); i++)
+for (i = 0; i < (int)strlen(s1); i++)
 concat[i] = s1[i];
 
-for (i = 0; i < strlen(s2); i++)
+for (i = 0; i < (int)strlen(s2); i++)
 concat[i + strlen(s1)] = s2[i];
 
 return (concat);
